apps/K9/pyproject.toml
[project]
name = "K9"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "bic",
    "holodeck",
    "captains_log",
]

[project.scripts]
K9 = "K9.cli:app"

[dependency-groups]
dev = ["pytest>=8", "respx>=0.21"]

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/K9"]

[tool.uv.sources]
bic = { workspace = true }
holodeck = { workspace = true }
captains_log = { workspace = true }
